nutanix/requirements.txt
# Core SpaceONE dependencies
spaceone-core==1.12.0
spaceone-api==1.12.0

# gRPC
grpcio==1.51.1
grpcio-tools==1.51.1
grpcio-health-checking==1.51.1

# HTTP client
httpx==0.24.0
aiohttp==3.8.3

# Logging
python-json-logger==2.0.5

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.2.2
pytest-asyncio==0.20.3
pytest-cov==4.1.0
httpx-mock==0.21.3

# Code quality
flake8==6.0.0
black==23.3.0
isort==5.12.0
mypy==1.0.1
